Delete FlexRow
This component was equivalent to a Box with display already set to flex and flexDirection already set to row.  This was cleaner than doing that common operation over and over using the sx prop but since I realized there is never any need for the sx prop on a Box component I realized this didn't actual save any complexity or clutter.