Move settings handlers to the modal.
Three pieces of state that only change at the beginning of a new game were amalgamated into a settings object. The settings modal maintains a local copy of the state complete with handlers for each type of setting change. This way the parent component is less cluttered. It has fewer state change handlers defined within it and fewer things it needs to pass to the modal.